% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_layout.R
\name{dashboard_layout}
\alias{dashboard_layout}
\title{Wrapper for dashboard layout}
\usage{
dashboard_layout(title, head = NULL, ...)
}
\arguments{
\item{...}{list of dashboard items. Use the dashboard_item function in order to create these items.}

\item{id}{shiny id}

\item{input}{internal}
}
\description{
Wrapper function for producing a dashboard layout. Each argument beyond the title
represents one dashboard tab. Use the dashboard_item function to prodcue each
tab.
}
\keyword{internal}
