\name{shiny_js}
\alias{shiny_js}
\title{Include a JavaScript File}
\usage{
  shiny_js(file)
}
\arguments{
  \item{file}{Location of the file.}
}
\description{
  This function produces a singleton for including a
  JavaScript file. Note that JavaScript files to be
  included in a Shiny server should be in the \code{www}
  folder; preferably \code{www/js}.
}

