\docType{package}
\name{shinyExtras}
\alias{shinyExtras}
\alias{shinyExtras-package}
\title{shinyExtras}
\description{
  This package contains a set of helper functions for R
  Shiny documents. Typically, these are functions that
  assist in the inclusion of JavaScript source / CSS style
  sheets, but we also include a handy HTML function for
  generation of HTML material from R code.
}

